cast htonl to unsigned int before printfing %u
git-svn-id: https://miau.svn.sourceforge.net/svnroot/miau/trunk@203 e7bde6ef-2016-0410-acfd-d3fa7a0704fe

+++ src/dcc.c
@@ -471,7 +471,8 @@ dcc_realinitiate(char *dest, dcccommand *dcc)
 
 	sprintf(dest, "\1DCC %s %s %u %u", dcc->type,
 			dcc->arg1,
-			ntohl(*(unsigned long int *) host->h_addr),
+			(unsigned int) ntohl(*(unsigned long int *)
+					host->h_addr),
 			myport);
 
 	if (dcc->argc == 5) {
